fix: work on conversion of mongo operators to sql

+++ src/libs/storage/framework/typeorm/TypeOrmSqlConditionsBuilder.ts
@@ -6,32 +6,52 @@ import {TypeOrmPropertyRef} from './schema/TypeOrmPropertyRef';
 import {IClassRef, IEntityRef} from 'commons-schema-api/browser';
 import {Brackets, EntityManager, QueryBuilder, SelectQueryBuilder} from 'typeorm';
 import {DateUtils} from 'typeorm/util/DateUtils';
+import {PAst} from '../../../expressions/ast/PAst';
+import {IMangoWalker} from '../../../expressions/IMangoWalker';
+import {And} from '../../../expressions/operators/logic/And';
+import {Or} from '../../../expressions/operators/logic/Or';
+import {AbstractCompare} from '../../../expressions/operators/compare/AbstractCompare';
+import {PValue} from '../../../expressions/ast/PValue';
+import {StorageRef} from '../../../../libs/storage/StorageRef';
+import {AbstractSchemaHandler} from '../../AbstractSchemaHandler';
+
+/**
+ * TODO Ugly build style make this better
+ */
+export class TypeOrmSqlConditionsBuilder<T> implements IMangoWalker {
 
 
-export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuilder*/ {
-
-
-  constructor(manager: EntityManager, entityRef: IEntityRef, type: 'update' | 'select' | 'delete', alias: string = null) {
+  constructor(manager: EntityManager | QueryBuilder<any>,
+              entityRef: IEntityRef,
+              storageRef: StorageRef,
+              type: 'update' | 'select' | 'delete',
+              alias: string = null) {
     // super(entityRef, alias);
     this.entityRef = entityRef;
     this.alias = alias;
     this.type = type;
-    const repo = manager.getRepository(entityRef.getClassRef().getClass());
+    this.baseStorageRef = storageRef;
+    this.handler = this.baseStorageRef.getSchemaHandler();
+    if (manager instanceof EntityManager) {
+      const repo = manager.getRepository(entityRef.getClassRef().getClass());
 
-    if (alias) {
-      this.baseQueryBuilder = repo.createQueryBuilder(alias) as QueryBuilder<T>;
+      if (alias) {
+        this.baseQueryBuilder = repo.createQueryBuilder(alias) as QueryBuilder<T>;
+      } else {
+        this.baseQueryBuilder = repo.createQueryBuilder() as QueryBuilder<T>;
+        this.alias = this.baseQueryBuilder.alias;
+      }
+
+      switch (type) {
+        case 'delete':
+          this.baseQueryBuilder = this.baseQueryBuilder.delete();
+          break;
+        case 'update':
+          this.baseQueryBuilder = this.baseQueryBuilder.update();
+          break;
+      }
     } else {
-      this.baseQueryBuilder = repo.createQueryBuilder() as QueryBuilder<T>;
-      this.alias = this.baseQueryBuilder.alias;
-    }
-
-    switch (type) {
-      case 'delete':
-        this.baseQueryBuilder = this.baseQueryBuilder.delete();
-        break;
-      case 'update':
-        this.baseQueryBuilder = this.baseQueryBuilder.update();
-        break;
+      this.baseQueryBuilder = manager;
     }
   }
 
@@ -48,11 +68,29 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
 
   type: 'update' | 'select' | 'delete' = 'select';
 
+  mode: 'where' | 'having' = 'where';
+
+  baseStorageRef: StorageRef;
+
+  handler: AbstractSchemaHandler;
+
   protected joins: IConditionJoin[] = [];
 
 
   getQueryBuilder(): QueryBuilder<T> {
     return this.baseQueryBuilder;
+  }
+
+  setQueryBuilder(qb: QueryBuilder<any>) {
+    this.baseQueryBuilder = qb;
+  }
+
+  getMode() {
+    return this.mode;
+  }
+
+  setMode(mode: 'where' | 'having') {
+    this.mode = mode;
   }
 
   lookupKeys(key: string): string {
@@ -61,6 +99,11 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
       case 'update':
         return key;
     }
+
+    if (this.mode === 'having') {
+      return key;
+    }
+
     const joins = key.split('.');
     let tmp: IClassRef = this.entityRef.getClassRef();
     let names: string[] = this.alias ? [this.alias] : [];
@@ -133,11 +176,24 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
     return names.join('.');
   }
 
+  private buildMango(condition: PAst) {
+    if (_.isEmpty(condition)) {
+      return null;
+    }
+    const brackets = condition.visit(this);
+    return this.asBracket(brackets);
+  }
+
 
   build(condition: any, k: string = null) {
     if (_.isEmpty(condition)) {
       return null;
     }
+
+    if (condition instanceof PAst) {
+      return this.buildMango(condition);
+    }
+
     // tslint:disable-next-line:no-shadowed-variable
     let control: any = _.keys(condition).filter(k => k.startsWith('$'));
     if (!_.isEmpty(control)) {
@@ -187,14 +243,31 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
   }
 
 
-  private _erg(op: string, condition: any, key: string = null, value: any = null) {
+  private _erg(op: string, key: string = null, value: any = null) {
     const _key = this.lookupKeys(key);
     const p = this.paramName();
     return new Brackets(qb => {
-      qb.where(`${_key} ${op} :${p}`, this.paramValue(p, value, _key));
+      if (this.mode === 'having') {
+        (<SelectQueryBuilder<any>>qb).having(`${_key} ${op} :${p}`, this.paramValue(p, value, _key));
+      } else {
+        qb.where(`${_key} ${op} :${p}`, this.paramValue(p, value, _key));
+      }
     });
   }
 
+  private handleOperation(op: string, key: string = null, value: any = null) {
+    const handle = this.handler.getOperationHandle(op.toLowerCase());
+    const _key = this.lookupKeys(key);
+    // const p = this.paramName();
+    return new Brackets(qb => {
+      const syntax = handle(_key, value);
+      if (this.mode === 'having') {
+        (<SelectQueryBuilder<any>>qb).having(syntax);
+      } else {
+        qb.where(syntax);
+      }
+    });
+  }
 
   private paramName() {
     return 'p' + (this.paramInc++);
@@ -211,7 +284,7 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
 
     // TODO make this more flexible
     if (v instanceof Date) {
-      if (ref) {
+      if (ref && this.mode === 'where') {
         const _columnName = columnName.split('.').pop();
         const entityMetadata = this.baseQueryBuilder.connection.getMetadata(ref.getClassRef().getClass());
         const columnMetadata = entityMetadata.columns.find(x => x.propertyName === _columnName);
@@ -228,31 +301,39 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
 
 
   $eq(condition: any, key: string = null, value: any = null) {
-    return this._erg('=', condition, key, value);
+    return this._erg('=', key, value);
   }
 
 
   $isNull(condition: any, key: string = null, value: any = null) {
     const _key = this.lookupKeys(key);
     return new Brackets(qb => {
-      qb.where(`${_key} IS NULL`);
+      if (this.mode === 'having') {
+        (<SelectQueryBuilder<any>>qb).having(`${_key} IS NULL`);
+      } else {
+        qb.where(`${_key} IS NULL`);
+      }
     });
   }
 
   $isNotNull(condition: any, key: string = null, value: any = null) {
     const _key = this.lookupKeys(key);
     return new Brackets(qb => {
-      qb.where(`${_key} IS NOT NULL`);
+      if (this.mode === 'having') {
+        (<SelectQueryBuilder<any>>qb).having(`${_key} IS NOT NULL`);
+      } else {
+        qb.where(`${_key} IS NOT NULL`);
+      }
     });
   }
 
   $ne(condition: any, key: string = null, value: any = null) {
-    return this._erg('<>', condition, key, value);
+    return this._erg('<>', key, value);
   }
 
 
   $lt(condition: any, key: string = null, value: any = null) {
-    return this._erg('<', condition, key, value);
+    return this._erg('<', key, value);
   }
 
   $lte(condition: any, key: string = null, value: any = null) {
@@ -260,11 +341,11 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
   }
 
   $le(condition: any, key: string = null, value: any = null) {
-    return this._erg('<=', condition, key, value);
+    return this._erg('<=', key, value);
   }
 
   $gt(condition: any, key: string = null, value: any = null) {
-    return this._erg('>', condition, key, value);
+    return this._erg('>', key, value);
   }
 
   $gte(condition: any, key: string = null, value: any = null) {
@@ -272,18 +353,29 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
   }
 
   $ge(condition: any, key: string = null, value: any = null) {
-    return this._erg('>=', condition, key, value);
+    return this._erg('>=', key, value);
   }
 
   $like(condition: any, key: string = null, value: any = null) {
-    return this._erg('LIKE', condition, key, value.replace(/%/g, '%%').replace(/\*/g, '%'));
+    return this._erg('LIKE', key, value.replace(/%/g, '%%').replace(/\*/g, '%'));
   }
+
+// TODO
+  $regex(condition: any, key: string = null, value: any = null) {
+    return this.handleOperation('regex', key, value);
+  }
+
+
 
   $in(condition: any, key: string = null, value: any = null) {
     const _key = this.lookupKeys(key);
     const p = this.paramName();
     return new Brackets(qb => {
-      qb.where(`${_key} IN (:...${p})`, this.paramValue(p, value));
+      if (this.mode === 'having') {
+        (<SelectQueryBuilder<any>>qb).having(`${_key} IN (:...${p})`, this.paramValue(p, value));
+      } else {
+        qb.where(`${_key} IN (:...${p})`, this.paramValue(p, value));
+      }
     });
   }
 
@@ -291,23 +383,128 @@ export class TypeOrmSqlConditionsBuilder<T> /*extends AbstractSqlConditionsBuild
     const _key = this.lookupKeys(key);
     const p = this.paramName();
     return new Brackets(qb => {
-      qb.where(`${_key} NOT IN (:...${p})`, this.paramValue(p, value));
+      if (this.mode === 'having') {
+        (<SelectQueryBuilder<any>>qb).having(`${_key} NOT IN (:...${p})`, this.paramValue(p, value));
+      } else {
+        qb.where(`${_key} NOT IN (:...${p})`, this.paramValue(p, value));
+      }
     });
+  }
+
+  // TODO
+  $not(condition: any, key: string = null, value: any = null) {
+    // return this.handleOperation('regex', key, value);
   }
 
   $and(condition: any, key: string = null) {
     return new Brackets(qb => {
-      _.map(condition['$and'], c => qb.andWhere(this.build(c, null)));
+      _.map(condition['$and'], c => {
+        if (this.mode === 'having') {
+          return (<SelectQueryBuilder<any>>qb).andHaving(this.build(c, null));
+        } else {
+          return qb.andWhere(this.build(c, null));
+        }
+      });
     });
-    // return '(' + ).join(') AND (') + ')';
   }
+
 
   $or(condition: any, key: string = null) {
     return new Brackets(qb => {
-      _.map(condition['$or'], c => qb.orWhere(this.build(c, null)));
+      _.map(condition['$or'], c => {
+        if (this.mode === 'having') {
+          return (<SelectQueryBuilder<any>>qb).orHaving(this.build(c, null));
+        } else {
+          return qb.orWhere(this.build(c, null));
+        }
+      });
     });
-    // return '(' + _.map(condition['$or'], c => this.build(c, null)).join(') OR (') + ')';
+  }
+
+  visitArray(ast: PAst): any[] {
+    return [];
+  }
+
+  leaveArray(brackets: any[], ast: PAst) {
+    return brackets;
+
+  }
+
+  visitObject(ast: PAst) {
+    return {};
+  }
+
+  leaveObject(res: any, ast: PAst) {
+    // return _.values(res);
+    return this.asBracket(_.values(res));
   }
 
 
+  onValue(ast: PAst): any {
+    if (ast instanceof PValue) {
+      return ast.value;
+    }
+    return null;
+  }
+
+
+  onOperator(ast: PAst, valueRes: any): any {
+    if (ast instanceof AbstractCompare) {
+      const method = '$' + ast.name;
+      if (this[method]) {
+        return this[method](null, ast.key, valueRes);
+      }
+    }
+    return null;
+  }
+
+  visitOperator(ast: PAst): any {
+    return {};
+  }
+
+  leaveOperator(res: any, ast: PAst) {
+    if (ast instanceof And) {
+      return new Brackets(qb => {
+        _.map(res, c => {
+          if (this.mode === 'having') {
+            return (<SelectQueryBuilder<any>>qb).andHaving(c);
+          } else {
+            return qb.andWhere(c);
+          }
+        });
+      });
+    } else if (ast instanceof Or) {
+      return new Brackets(qb => {
+        _.map(res, c => {
+          if (this.mode === 'having') {
+            return (<SelectQueryBuilder<any>>qb).orHaving(c);
+          } else {
+            return qb.orWhere(c);
+          }
+        });
+      });
+    }
+    return res;
+  }
+
+
+  asBracket(brackets: any) {
+    if (_.isArray(brackets)) {
+      if (brackets.length === 1) {
+        return _.first(brackets);
+      } else if (brackets.length > 1) {
+        return new Brackets(qb => {
+          _.map(brackets, c => {
+            if (this.mode === 'having') {
+              return (<SelectQueryBuilder<any>>qb).andHaving(c);
+            } else {
+              return qb.andWhere(c);
+            }
+          });
+        });
+      }
+    }
+    return brackets;
+
+  }
 }
